Moved helpers to common.rs and introduced various input sizes
